requirements.txt
requests>=2.28.2
openai>=0.27.2
python-dotenv>=1.0.0
rich>=13.3.5
